Historial Mascota Terminado, a  revision de Front

+++ BackPuppy/Dtos/ControlFisicoDtoOut.cs
@@ -4,7 +4,7 @@ namespace BackPuppy.Dtos
 {
     public class ControlFisicoDtoOut
     {
-        public BigInteger id_control_fisico { get; set; }
+        public int id_control_fisico { get; set; }
         public string? temperatura { get; set; }
         public string? frecCardiaca { get; set; }
         public string? frecRespiratoria { get; set; }
